+++ src/main/java/com/telegram/drb/repository/impl/mapper/UserReputationMapper.java
@@ -17,7 +17,7 @@ public class UserReputationMapper implements RowMapper<UserReputation> {
     public UserReputation mapRow(ResultSet rs, int rowNum) throws SQLException {
         UserReputation userReputation = new UserReputation();
         userReputation.setUserId(rs.getInt("user_id"));
-        userReputation.setChatId(rs.getInt("chat_id"));
+        userReputation.setChatId(rs.getLong("chat_id"));
         userReputation.setReputationValue(rs.getInt("reputation_value"));
         return userReputation;
     }
